Remove the download panel, move auto download to StudyEnd
